// packages/server/interfaces/IHeaderLabelRetorno.ts
export interface IHeaderLabelRetorno {
  idRegistro: string;
  idRetorno: string;
  literalRetorno: string;
  codigoServico: string;
  literalServico: string;
  codigoEmpresa: string;
  nomeEmpresa: string;
  numeroBradesco: string;
  nomeBanco: string;
  dataArquivo: string;
  densidadeGravacao: string;
  numeroAvisoBancario: string;
  dataCredito: string;
  sequenciaRegistro: number;
}