// packages/server/interfaces/ITraillerRetorno.ts
export interface ITraillerRetorno {
  idRegistro: string;
  idRetorno: string;
  tipoRegistro: string;
  codigoBanco: string;
  quantidadeTitulos: string;
  valorTotal: string;
  numeroAvisoBancario: string;
  quantidadeRegistros02: string;
  valorRegistros02: string;
  valorRegistros: string;
  quantidadeRegistros06: string;
  valorRegistros06: string;
  quantidadeRegistrosBaixados: string;
  valorRegistrosBaixados: string;
  quantidadeRegistrosAbatimentoCancelado: string;
  valorRegistrosAbatimentoCancelado: string;
  quantidadeRegistrosVencimentoAlterado: string;
  valorRegistrosVencimentoAlterado: string;
  quantidadeRegistrosAbatimento: string;
  valorRegistrosAbatimento: string;
  quantidadeRegistrosConfirmacaoProtesto: string;
  valorRegistrosConfirmacaoProtesto: string;
  valorTotalRateios: string;
  quantidadeRateios: string;
  sequenciaRegistro: string;
}